+++ data/index.js
@@ -1,17 +1,25 @@
+const makeChainFeeRateResponse = require('./make_chain_fee_rate_response');
 const makeForwardsResponse = require('./make_forwards_response');
+const makeFundPsbtResponse = require('./make_fund_psbt_response');
 const makeInvoice = require('./make_invoice');
 const makePayViaRoutesResponse = require('./make_pay_via_routes_response');
 const makePaymentData = require('./make_payment_data');
 const makePaymentRequest = require('./make_payment_request');
 const makeRoutesResponse = require('./make_routes_response');
+const makeSignPsbtResponse = require('./make_sign_psbt_response');
 const makeWalletInfoResponse = require('./make_wallet_info_response');
+const makeWalletVersionResponse = require('./make_wallet_version_response');
 
 module.exports = {
+  makeChainFeeRateResponse,
   makeForwardsResponse,
+  makeFundPsbtResponse,
   makeInvoice,
   makePayViaRoutesResponse,
   makePaymentData,
   makePaymentRequest,
   makeRoutesResponse,
+  makeSignPsbtResponse,
   makeWalletInfoResponse,
+  makeWalletVersionResponse,
 };